Set up new system for interaction implementations.
This doesn't build yet, but since there's a lot of work I'm making a
commit. The changes here begin the macro system for autogenerating
interaction types, methods, and their implementations, with escape
hatches for special behavior in the add case.
