+++ deepwell/src/services/interaction/structs.rs
@@ -133,6 +133,7 @@ pub enum InteractionType {
     PageWatch,
     UserFollow,
     UserContact,
+    UserContactRequest,
     UserBlock,
 }
 
@@ -149,6 +150,7 @@ impl InteractionType {
             InteractionType::PageWatch => "watch",
             InteractionType::UserFollow => "follow",
             InteractionType::UserContact => "contact",
+            InteractionType::UserContactRequest => "contact-request",
             InteractionType::UserBlock => "block",
         }
     }
@@ -170,32 +172,8 @@ impl InteractionType {
             InteractionType::PageWatch => t!(Page, User),
             InteractionType::UserFollow => t!(User, User),
             InteractionType::UserContact => t!(User, User),
+            InteractionType::UserContactRequest => t!(User, User),
             InteractionType::UserBlock => t!(User, User),
         }
     }
-}
-
-// Per-interaction data structures
-//
-// If you change these remember you need to account for backwards compatibility!
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub struct SiteBanData {
-    pub banned_until: Option<Date>,
-    pub reason: String,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub struct SiteMemberData {
-    pub accepted: SiteMemberAccepted,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-#[serde(rename_all = "snake_case", tag = "cause", content = "user_id")]
-pub enum SiteMemberAccepted {
-    CreatedSite,
-    SelfJoined,
-    Password,
-    Accepted(i64),
-    Invitation(i64),
 }
